+++ software/o_c_REV/HEM_ClockSetup.ino
@@ -61,15 +61,15 @@ public:
         }
     }
 
-    uint32_t OnDataRequest() {
-        uint32_t data = 0;
+    uint64_t OnDataRequest() {
+        uint64_t data = 0;
         Pack(data, PackLocation { 0, 1 }, clock_m->IsRunning() || clock_m->IsPaused());
         Pack(data, PackLocation { 1, 9 }, clock_m->GetTempo());
         Pack(data, PackLocation { 10, 5 }, clock_m->GetMultiply());
         return data;
     }
 
-    void OnDataReceive(uint32_t data) {
+    void OnDataReceive(uint64_t data) {
         if (Unpack(data, PackLocation { 0, 1 })) {
             clock_m->Start();
             clock_m->Pause();
@@ -149,5 +149,5 @@ void ClockSetup_View(bool hemisphere) {ClockSetup_instance[hemisphere].BaseView(
 void ClockSetup_OnButtonPress(bool hemisphere) {ClockSetup_instance[hemisphere].OnButtonPress();}
 void ClockSetup_OnEncoderMove(bool hemisphere, int direction) {ClockSetup_instance[hemisphere].OnEncoderMove(direction);}
 void ClockSetup_ToggleHelpScreen(bool hemisphere) {ClockSetup_instance[hemisphere].HelpScreen();}
-uint32_t ClockSetup_OnDataRequest(bool hemisphere) {return ClockSetup_instance[hemisphere].OnDataRequest();}
-void ClockSetup_OnDataReceive(bool hemisphere, uint32_t data) {ClockSetup_instance[hemisphere].OnDataReceive(data);}
+uint64_t ClockSetup_OnDataRequest(bool hemisphere) {return ClockSetup_instance[hemisphere].OnDataRequest();}
+void ClockSetup_OnDataReceive(bool hemisphere, uint64_t data) {ClockSetup_instance[hemisphere].OnDataReceive(data);}
